+++ crates/nostr/examples/nip09.rs
@@ -14,7 +14,7 @@ fn main() -> Result<()> {
         .id(event_id)
         .reason("these posts were published by accident");
 
-    let event: Event = EventBuilder::delete(request).sign_with_keys(&keys)?;
+    let event: Event = EventBuilder::delete(request).sign(&keys)?;
     println!("{}", event.as_json());
 
     Ok(())