+++ crates/nostr/src/event/mod.rs
@@ -383,7 +383,7 @@ mod tests {
     fn test_custom_kind() {
         let keys = Keys::generate();
         let e: Event = EventBuilder::new(Kind::Custom(123), "my content")
-            .sign_with_keys(&keys)
+            .sign(&keys)
             .unwrap();
 
         let serialized = e.as_json();
@@ -400,7 +400,7 @@ mod tests {
         let my_keys = Keys::generate();
         let event = EventBuilder::text_note("my content")
             .tags([Tag::expiration(Timestamp::from(1600000000))])
-            .sign_with_keys(&my_keys)
+            .sign(&my_keys)
             .unwrap();
 
         assert!(&event.is_expired());
@@ -415,7 +415,7 @@ mod tests {
         let my_keys = Keys::generate();
         let event = EventBuilder::text_note("my content")
             .tags([Tag::expiration(Timestamp::from(expiry_date))])
-            .sign_with_keys(&my_keys)
+            .sign(&my_keys)
             .unwrap();
 
         assert!(!&event.is_expired());
@@ -426,7 +426,7 @@ mod tests {
     fn test_event_without_expiration_tag() {
         let my_keys = Keys::generate();
         let event = EventBuilder::text_note("my content")
-            .sign_with_keys(&my_keys)
+            .sign(&my_keys)
             .unwrap();
         assert!(!&event.is_expired());
     }