+++ crates/nostr/src/nips/nip09.rs
@@ -116,7 +116,7 @@ mod tests {
             .coordinate(coordinate)
             .reason("these posts were published by accident");
 
-        let event: Event = request.to_event_builder().sign_with_keys(&keys).unwrap();
+        let event: Event = request.to_event_builder().sign(&keys).unwrap();
 
         assert_eq!(event.kind, Kind::EventDeletion);
         assert_eq!(event.content, "these posts were published by accident");
@@ -137,7 +137,7 @@ mod tests {
         // Event ID without reason
         let request = EventDeletionRequest::new().id(event_id);
 
-        let event: Event = request.to_event_builder().sign_with_keys(&keys).unwrap();
+        let event: Event = request.to_event_builder().sign(&keys).unwrap();
 
         assert_eq!(event.kind, Kind::EventDeletion);
         assert!(event.content.is_empty());