+++ crates/nostr/src/nips/nip34.rs
@@ -949,11 +949,7 @@ mod tests {
         };
 
         let keys = Keys::generate();
-        let event: Event = repo
-            .to_event_builder()
-            .unwrap()
-            .sign_with_keys(&keys)
-            .unwrap();
+        let event: Event = repo.to_event_builder().unwrap().sign(&keys).unwrap();
 
         assert_eq!(event.kind, Kind::GitRepoAnnouncement);
         assert!(event.content.is_empty());
@@ -1013,11 +1009,7 @@ mod tests {
         };
 
         let keys = Keys::generate();
-        let event: Event = repo
-            .to_event_builder()
-            .unwrap()
-            .sign_with_keys(&keys)
-            .unwrap();
+        let event: Event = repo.to_event_builder().unwrap().sign(&keys).unwrap();
 
         assert_eq!(event.kind, Kind::GitIssue);
         assert_eq!(event.content, "My issue content");
@@ -1065,11 +1057,7 @@ mod tests {
         };
 
         let keys = Keys::generate();
-        let event: Event = repo
-            .to_event_builder()
-            .unwrap()
-            .sign_with_keys(&keys)
-            .unwrap();
+        let event: Event = repo.to_event_builder().unwrap().sign(&keys).unwrap();
 
         assert_eq!(event.kind, Kind::GitPatch);
         assert_eq!(event.content, "<patch>");
@@ -1126,11 +1114,7 @@ mod tests {
         };
 
         let keys = Keys::generate();
-        let event: Event = update
-            .to_event_builder()
-            .unwrap()
-            .sign_with_keys(&keys)
-            .unwrap();
+        let event: Event = update.to_event_builder().unwrap().sign(&keys).unwrap();
 
         assert_eq!(event.kind, Kind::GitPullRequestUpdate);
         assert!(event.content.is_empty());