+++ crates/nostr/src/nips/nip42.rs
@@ -167,7 +167,7 @@ mod tests {
         let challenge = "1234567890";
 
         let event = EventBuilder::auth(challenge, relay_url.clone())
-            .sign_with_keys(&keys)
+            .sign(&keys)
             .unwrap();
 
         assert!(is_valid_auth_event(&event, &relay_url, challenge));
@@ -181,20 +181,18 @@ mod tests {
 
         // Wrong challenge
         let event = EventBuilder::auth("abcd", relay_url.clone())
-            .sign_with_keys(&keys)
+            .sign(&keys)
             .unwrap();
         assert!(!is_valid_auth_event(&event, &relay_url, challenge));
 
         // Wrong relay url
         let event = EventBuilder::auth(challenge, RelayUrl::parse("wss://example.com").unwrap())
-            .sign_with_keys(&keys)
+            .sign(&keys)
             .unwrap();
         assert!(!is_valid_auth_event(&event, &relay_url, challenge));
 
         // Wrong kind
-        let event = EventBuilder::text_note("abcd")
-            .sign_with_keys(&keys)
-            .unwrap();
+        let event = EventBuilder::text_note("abcd").sign(&keys).unwrap();
         assert!(!is_valid_auth_event(&event, &relay_url, challenge));
     }
 }