+++ sdk/src/relay/api/sync.rs
@@ -647,13 +647,13 @@ mod tests {
         // Build events to store in the local database
         let local_events = [
             EventBuilder::text_note("Local 1")
-                .sign_with_keys(&Keys::generate())
+                .sign(&Keys::generate())
                 .unwrap(),
             EventBuilder::text_note("Local 2")
-                .sign_with_keys(&Keys::generate())
+                .sign(&Keys::generate())
                 .unwrap(),
             EventBuilder::new(Kind::Custom(123), "Local 123")
-                .sign_with_keys(&Keys::generate())
+                .sign(&Keys::generate())
                 .unwrap(),
         ];
 
@@ -678,13 +678,13 @@ mod tests {
             // Event in common with the local database
             local_events[0].clone(),
             EventBuilder::text_note("Test 2")
-                .sign_with_keys(&Keys::generate())
+                .sign(&Keys::generate())
                 .unwrap(),
             EventBuilder::text_note("Test 3")
-                .sign_with_keys(&Keys::generate())
+                .sign(&Keys::generate())
                 .unwrap(),
             EventBuilder::new(Kind::Custom(123), "Test 4")
-                .sign_with_keys(&Keys::generate())
+                .sign(&Keys::generate())
                 .unwrap(),
         ];