+++ crates/nostr/CHANGELOG.md
@@ -45,6 +45,7 @@
 - Remove TagKind (https://github.com/rust-nostr/nostr/pull/1347)
 - Drop support for private and anon zaps (https://github.com/rust-nostr/nostr/pull/1355)
 - Remove `EventBuilder::sign_with_keys` and `EventBuilder::sign_with_ctx` (https://github.com/rust-nostr/nostr/pull/1355)
+- Remove `UnsignedEvent::sign_with_keys`, `UnsignedEvent::sign_with_ctx` and `UnsignedEvent::sign_with_aux_rand` (https://github.com/rust-nostr/nostr/pull/1355)
 
 ### Changed